requirements.txt
asgiref==3.7.2
astroid==2.15.6
certifi==2023.7.22
charset-normalizer==3.2.0
dill==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
graphviz==0.20.1
gunicorn==21.2.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.4
mccabe==0.7.0
packaging==23.1
platformdirs==3.10.0
psycopg2-binary==2.9.7
pylint==2.17.5
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
wrapt==1.15.0
